Rename `__generated__` to `__gen__` in web app
