+++ apps/web/src/components/Day.tsx
@@ -1,15 +1,15 @@
 import { useEffect, useRef, useState } from "react";
 import { graphql, useFragment, useMutation } from "@flowdev/relay";
-import { Day_day$key } from "@flowdev/web/relay/__generated__/Day_day.graphql";
+import { Day_day$key } from "@flowdev/web/relay/__gen__/Day_day.graphql";
 import { TaskCard } from "./TaskCard";
 import { dayjs } from "@flowdev/web/dayjs";
 import { ReactSortable, Sortable } from "react-sortablejs";
-import { DayContent_day$key } from "@flowdev/web/relay/__generated__/DayContent_day.graphql";
-import { DayUpdateTaskDateMutation } from "@flowdev/web/relay/__generated__/DayUpdateTaskDateMutation.graphql";
-import { DayAddTaskActionsBar_day$key } from "@flowdev/web/relay/__generated__/DayAddTaskActionsBar_day.graphql";
+import { DayContent_day$key } from "@flowdev/web/relay/__gen__/DayContent_day.graphql";
+import { DayUpdateTaskDateMutation } from "@flowdev/web/relay/__gen__/DayUpdateTaskDateMutation.graphql";
+import { DayAddTaskActionsBar_day$key } from "@flowdev/web/relay/__gen__/DayAddTaskActionsBar_day.graphql";
 import { Button } from "@flowdev/ui/Button";
 import { environment } from "@flowdev/web/relay/environment";
-import { OnCreateTaskItemRecordToCreateTaskFrom_item$data } from "@flowdev/web/relay/__generated__/OnCreateTaskItemRecordToCreateTaskFrom_item.graphql";
+import { OnCreateTaskItemRecordToCreateTaskFrom_item$data } from "@flowdev/web/relay/__gen__/OnCreateTaskItemRecordToCreateTaskFrom_item.graphql";
 import { getPlugins } from "@flowdev/web/getPlugin";
 import { OnCreateTask, OnCreateTaskProps, PluginStepInfo } from "./OnCreateTask";
 import { DragContext, useDragContext } from "../useDragContext";
