+++ apps/web/src/components/ItemTitle.tsx
@@ -2,9 +2,9 @@ import { useCallback, useEffect, useRef, useState } from "react";
 import { Editor, EditorContent, useEditor, Mention, MinimumKit, OnEscape } from "@flowdev/tiptap";
 import { CatchNewLines } from "@flowdev/tiptap";
 import { graphql, useFragment, useMutation } from "@flowdev/relay";
-import { ItemTitle_item$key } from "../relay/__generated__/ItemTitle_item.graphql";
-import { ItemTitleUpdateItemTitleMutation } from "../relay/__generated__/ItemTitleUpdateItemTitleMutation.graphql";
-import { ItemTitleCreateItemMutation } from "../relay/__generated__/ItemTitleCreateItemMutation.graphql";
+import { ItemTitle_item$key } from "../relay/__gen__/ItemTitle_item.graphql";
+import { ItemTitleUpdateItemTitleMutation } from "../relay/__gen__/ItemTitleUpdateItemTitleMutation.graphql";
+import { ItemTitleCreateItemMutation } from "../relay/__gen__/ItemTitleCreateItemMutation.graphql";
 import { deleteVirtualItem, isTempItemId } from "./InboxList";
 import "./TaskTitle.scss";
