+++ apps/web/src/getPlugin/createTask.ts
@@ -1,6 +1,6 @@
 import { commitMutationPromise, graphql } from "@flowdev/relay";
 import { environment } from "../relay/environment";
-import { createTaskMutation } from "../relay/__generated__/createTaskMutation.graphql";
+import { createTaskMutation } from "../relay/__gen__/createTaskMutation.graphql";
 
 export const createTask = async (input: createTaskMutation["variables"]["input"]) => {
   const { createTask } = await commitMutationPromise<createTaskMutation>(environment, {
