+++ apps/web/src/views/BrowsePluginsView.tsx
@@ -6,8 +6,8 @@ import { Popover, PopoverContent, PopoverTrigger } from "@flowdev/ui/Popover";
 import { useForm } from "react-hook-form";
 import { graphql, useMutation } from "@flowdev/relay";
 import { Button } from "@flowdev/ui/Button";
-import { BrowsePluginsViewInstallMutation } from "../relay/__generated__/BrowsePluginsViewInstallMutation.graphql";
-import { BrowsePluginsViewInstallFromUrlMutation } from "@flowdev/web/relay/__generated__/BrowsePluginsViewInstallFromUrlMutation.graphql";
+import { BrowsePluginsViewInstallMutation } from "../relay/__gen__/BrowsePluginsViewInstallMutation.graphql";
+import { BrowsePluginsViewInstallFromUrlMutation } from "@flowdev/web/relay/__gen__/BrowsePluginsViewInstallFromUrlMutation.graphql";
 import { useState } from "react";
 import { toast } from "@flowdev/ui/Toast";
 import { usePlugins } from "../getPlugin";
