+++ apps/web/vite.config.ts
@@ -26,7 +26,7 @@ const relay = () =>
 
           const graphqlFileName = `graphql__${crypto.randomBytes(10).toString("hex")}`;
           imports.push(
-            `import ${graphqlFileName} from "@flowdev/web/relay/__generated__/${operationName}.graphql.ts";`,
+            `import ${graphqlFileName} from "@flowdev/web/relay/__gen__/${operationName}.graphql.ts";`,
           );
           return graphqlFileName;
         });
